categories support users

+++ src/main/java/ru/ist/category/CategoryServiceImpl.java
@@ -7,9 +7,11 @@ import ru.ist.category.dto.CategoryDto;
 import ru.ist.category.dto.CategoryInputDto;
 import ru.ist.category.dto.CategoryUpdateDto;
 import ru.ist.category.model.Category;
+import ru.ist.error.exceptions.Forbidden;
 import ru.ist.operation.model.OperationType;
 import ru.ist.service.MapperService;
 import ru.ist.service.ValidationService;
+import ru.ist.user.model.User;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -23,37 +25,52 @@ public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
 
     @Override
-    public CategoryDto createCategory(CategoryInputDto categoryInputDto) {
-        Category category = mapperService.toCategory(categoryInputDto);
+    public CategoryDto createCategory(Long userId, CategoryInputDto categoryInputDto) {
+        User user = validationService.validateUser(userId);
+        Category category = categoryRepository.save(mapperService.toCategory(categoryInputDto, user));
 
         log.info("Создана новая категория: " + category);
-        return mapperService.toCategoryDto(categoryRepository.save(category));
+        return mapperService.toCategoryDto(category);
     }
 
     @Override
-    public List<CategoryDto> getCategories(OperationType operationType) {
+    public List<CategoryDto> getCategories(Long userId, OperationType operationType) {
+        User user = validationService.validateUser(userId);
+
         if (operationType != null) {
-            return categoryRepository.findByOperationType(operationType).stream()
+            log.info("Получена информация о категориях по типу операции");
+            return categoryRepository.findByUser_IdAndOperationType(userId, operationType).stream()
                     .map(mapperService::toCategoryDto)
                     .collect(Collectors.toList());
         }
 
-        return categoryRepository.findAll().stream()
+        log.info("Получена информация обо всех категориях");
+        return categoryRepository.findByUser_Id(userId).stream()
                 .map(mapperService::toCategoryDto)
                 .collect(Collectors.toList());
     }
 
     @Override
-    public CategoryDto getCategoryById(Long id) {
+    public CategoryDto getCategoryById(Long userId, Long id) {
+        User user = validationService.validateUser(userId);
         Category category = validationService.validateCategory(id);
+
+        if (!category.getUser().getId().equals(userId)) {
+            throw new Forbidden();
+        }
 
         log.info("Получение информации о категории: " + category);
         return mapperService.toCategoryDto(category);
     }
 
     @Override
-    public CategoryDto updateCategoryById(Long id, CategoryUpdateDto categoryUpdateDto) {
+    public CategoryDto updateCategoryById(Long userId, Long id, CategoryUpdateDto categoryUpdateDto) {
+        User user = validationService.validateUser(userId);
         Category category = validationService.validateCategory(id);
+
+        if (!category.getUser().getId().equals(userId)) {
+            throw new Forbidden();
+        }
 
         category.setTitle(categoryUpdateDto.getTitle() == null ? category.getTitle() : categoryUpdateDto.getTitle());
 
@@ -62,8 +79,13 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategoryById(Long id) {
-        validationService.validateCategory(id);
+    public void deleteCategoryById(Long userId, Long id) {
+        User user = validationService.validateUser(userId);
+        Category category = validationService.validateCategory(id);
+
+        if (!category.getUser().getId().equals(userId)) {
+            throw new Forbidden();
+        }
 
         log.info("Категория с id = " + id + " успешно удалена");
         categoryRepository.deleteById(id);
